+++ src/main/java/com/sivalabs/jblogger/repositories/UserRepository.java
@@ -7,12 +7,8 @@ import com.sivalabs.jblogger.entities.User;
 
 import java.util.Optional;
 
-/**
- * @author Siva
- * 
- */
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer>
+public interface UserRepository extends JpaRepository<User, Long>
 {
 	Optional<User> findByEmail(String email);
 }
